Make sure the function exists before setting the install date. Should only be another load or two later.

+++ inc/base.php
@@ -4,7 +4,7 @@ function mm_setup() {
 	if ( ( '' === get_option( 'mm_master_aff' ) || false === get_option( 'mm_master_aff' ) ) && defined( 'MMAFF' ) ) {
 		update_option( 'mm_master_aff', MMAFF );
 	}
-	if ( ! get_option( 'mm_install_date' ) ) {
+	if ( ! get_option( 'mm_install_date' ) && function_exists( 'save_mod_rewrite_rules' ) ) {
 		update_option( 'mm_install_date', date( 'M d, Y' ) );
 		$event = array(
 			't'		=> 'event',
